mapa3D/requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.8.0
rasterio>=1.3.0
GDAL>=3.4.0
numpy-stl>=3.0.0
scikit-image>=0.19.0
meshio>=5.3.0
trimesh>=3.15.0
geopandas>=0.12.0
shapely>=1.8.0
PyQt5>=5.15.0
Pillow>=9.0.0
tqdm>=4.64.0
pytest>=7.0.0
flake8>=4.0.0
